Move CSharpHighlightingTokenWriter to outer class
